Refactor rule converter
