+++ tools/convert/helpers_test.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"reflect"
 	"testing"
 )
 
@@ -25,26 +24,6 @@ func Test_conditional(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			if got := conditional(tt.data, tt.key, tt.extra); got != tt.want {
 				t.Errorf("conditional() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_keysToLowercase(t *testing.T) {
-	tests := []struct {
-		name string
-		m    map[string]any
-		want map[string]any
-	}{
-		{"empty", map[string]any{}, map[string]any{}},
-		{"one", map[string]any{"A": "b"}, map[string]any{"a": "b"}},
-		{"two", map[string]any{"A": "b", "C": "d"}, map[string]any{"a": "b", "c": "d"}},
-		{"recursive", map[string]any{"A": "b", "C": map[string]any{"D": "e"}}, map[string]any{"a": "b", "c": map[string]any{"d": "e"}}},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := _keysToLowercase(tt.m); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("_keysToLowercase() = %v, want %v", got, tt.want)
 			}
 		})
 	}
